image/test/mochitest/test_bug1359291.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1359291
-->
<head>
  <title>Test for Bug 1359291</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <script type="application/javascript" src="imgutils.js"></script>
  <script type="application/javascript" src="animationPolling.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script type="text/javascript">
const FAILURE_TIMEOUT = 60000; // Fail early after 60 seconds

SimpleTest.waitForExplicitFinish();

var expectedLoads = 1;
var expectedErrors = 8;
var observedLoads = 0;
var observedErrors = 0;

function getPath()
{
  return "bug1359291.sjs?" + (observedErrors == expectedErrors ? "true" : "false");
}

function onLoad(img)
{
  ++observedLoads;
  ok(observedLoads <= expectedLoads, "current onLoad " + observedLoads + " of " + expectedLoads);
  ok(observedErrors == expectedErrors, "previously observed correct number on onErrors");
  img.style = "display:none; width: 100px; height: 100px;";

  var animTest = new AnimationTest(20, FAILURE_TIMEOUT, 'referenceDiv',
                                   'animatedGif', 'debug');
  animTest.beginTest();
}

function onError(img)
{
  ++observedErrors;
  ok(observedErrors <= expectedErrors, "onError " + observedErrors + " of " + expectedErrors);
  ok(observedLoads == 0, "should not have observed any previous onLoads");
  img.src = getPath();
}
</script>
</pre>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=641748">
Mozilla Bug 1359291: Image loading errors can cause flickering.
</a>
<p id="display"></p>
<div id="content">
  <div id="referenceDiv" style="height: 100px; width: 100px;
                                display: none; background: #0018ff;"></div>
  <div id="animatedImage">
    <img id="animatedGif" src="bug1359291.sjs?false" onerror="onError(this)" onload="onLoad(this)"/>
    <div id="text-descr"></div>
    </div>
    <div id="debug" style="display:none">
  </div>
</div>
</body>
</html>
